Support systemd
Adds a service unit to control Minecraft with systemd.

Also adds a socket unit which controls a named pipe; systemd will attach
this to the server's standard input. This allows us to control the
server by writing to the named pipe.

This is a non-breaking change; Supervisor remains the default process
controller.