Chain handlers to achieve conditional behaviour
Variable substitution only works in task or include names when the
variable has been passed in as a role parameter.

This means using `{{ minecraft_process_control }}` in task names to
achieve dynamic behaviour will not work if the user is using the default
value of `minecraft_process_control`.

Instead, we define a generic handler "entry point" and chain it to the
process supervisor-specific actions.