Test server.properties modifications.
Set the rcon port and password, and verify that the port is open.

I had to bump up the minecraft-boot timeout from 5 seconds to 10, as 5
seconds was, interestingly, enough time for the main port to open but not
always enough time for the rcon port to open.  Ideally, the sleep should be
replaced by a retry loop that only runs the tests when we verify that
Minecraft is fully up (or after it times out).