Add ServerProperties class to minecraft_server_file module.
ServerProperties manages the server.properties file.  It differs from the
ACL files in that it is not JSON but rather a simple list of variables
in a "name=value" format, similar to an INI file.  These are provided as a
dict, which is passed to the `values` argument, which in turn has its type
enforced by the subclass of ServerFile.